+++ ui/swing/tools/src/main/java/imagej/ui/swing/tools/overlay/EllipseAdapter.java
@@ -64,7 +64,7 @@ public class EllipseAdapter extends
 	AbstractJHotDrawOverlayAdapter<EllipseOverlay>
 {
 
-	public static final int PRIORITY = RectangleAdapter.PRIORITY - 1;
+	public static final double PRIORITY = RectangleAdapter.PRIORITY - 1;
 
 	static protected EllipseOverlay downcastOverlay(final Overlay roi) {
 		assert (roi instanceof EllipseOverlay);
