+++ ui/swing/tools/src/main/java/imagej/ui/swing/tools/overlay/PolygonAdapter.java
@@ -76,7 +76,7 @@ public class PolygonAdapter extends
 	AbstractJHotDrawOverlayAdapter<PolygonOverlay>
 {
 
-	public static final int PRIORITY = EllipseAdapter.PRIORITY - 1;
+	public static final double PRIORITY = EllipseAdapter.PRIORITY - 1;
 
 	static private BezierFigure downcastFigure(final Figure figure) {
 		assert figure instanceof BezierFigure;
